Updates rollup configurations
