+++ src/index.spec.ts
@@ -1,15 +1,12 @@
 import {
-  extractAaguid,
-  getAaguid,
-  base64UrlToBase64,
-  base64ToUint8Array,
   findAuthenticatorById,
   getAuthenticator,
   getAuthenticatorId,
   getAuthenticatorName,
-} from '../index'
+} from './index'
 
-import { authenticatorNamesList } from '../authenticator-names'
+import { authenticatorNamesList } from './authenticator-names'
+import { base64UrlToBase64, base64ToUint8Array, getAaguid, extractAaguid } from './lib';
 
 const authenticatorDataBase64Url =
   "SZYN5YgOjGh0NBcPZHZgW4_krrmihjLHmVzzuoMdl2NdAAAAAPv8MAcVTk7MjAtuAgVX170AFBUCgIutOmnd-P3TTsakYoMM292opQECAyYgASFYIGTgM0IiDgO9AqTMSMT1Tdh1sHiL99qEZJ4cdk8vJAyDIlggolBgLLxO9I2q9GuYsa8kBThr8-iXpiO4mL2z_73-Th4";
